[CIVP-13473] Git clone with error handling(#17)
* removing git flag from post_save

* removing git-clone script

* create directories if not present

* cloning done completely via GitPython

* error with leading slashes

* startup script for cloning

* logging utilities

* creating log folder/file if necessary

* drying the code

* fixing flake8

* skipping set up if error log file has logs

* setting redirect based on git installation

* python2&3 compatible and making paths consistent

* fixing paths

* add a global gitignore file

* fixing broken unit tests

* renaming error class

* DRYing the code

* forgot setup.py

+++ civis_jupyter_notebooks/assets/jupyter_notebook_config.py
@@ -8,47 +8,48 @@ them not being reimported.
 import os
 import signal
 import pip
-from civis_jupyter_notebooks import platform_persistence
+from civis_jupyter_notebooks import platform_persistence, log_utils
 from civis_jupyter_notebooks.platform_persistence import NotebookManagementError
 
-git_enabled = os.environ.get('GIT_FILE') is not None
-
-if git_enabled:
-    NOTEBOOK_PATH = os.path.join('~', 'work', os.environ.get("GIT_FILE"))
-else:
-    NOTEBOOK_PATH = os.path.join('~', 'work', 'notebook.ipynb')
-NOTEBOOK_PATH = os.path.expanduser(NOTEBOOK_PATH)
-
+# Jupyter Configuration
 c = get_config() # noqa
-
 c.NotebookApp.ip = '*'
 c.NotebookApp.allow_origin = '*'
 c.NotebookApp.port = 8888
 c.NotebookApp.open_browser = False
 c.NotebookApp.token = ''
 c.NotebookApp.tornado_settings = {'headers': {'Content-Security-Policy': "frame-ancestors *"}}
-c.MultiKernelManager.default_kernel_name = os.environ['DEFAULT_KERNEL']
 c.NotebookApp.allow_root = True
+c.FileContentsManager.post_save_hook = platform_persistence.post_save
+c.MultiKernelManager.default_kernel_name = os.environ['DEFAULT_KERNEL']
 
-# Download notebook and initialize post-save hook
-try:
-    if not git_enabled or not os.path.isfile(NOTEBOOK_PATH):
-        platform_persistence.initialize_notebook_from_platform(NOTEBOOK_PATH)
-    # force a save of the preview so that we have one in case
-    # the user never generates one
-    _, preview_url = platform_persistence.get_update_urls()
-    platform_persistence.generate_and_save_preview(preview_url, NOTEBOOK_PATH)
-except NotebookManagementError as e:
-    platform_persistence.logger.error(str(e))
-    platform_persistence.logger.warn('Killing the notebook process b/c of a startup issue')
-    os.kill(os.getpid(), signal.SIGTERM)
+ROOT_DIR = os.path.expanduser(os.path.join('~', 'work'))
 
-post_save = platform_persistence.post_save(git_enabled=git_enabled)
-c.FileContentsManager.post_save_hook = post_save
+if log_utils.log_file_has_logs(log_utils.USER_VISIBLE_LOGS):
+    # redirect to log file
+    c.NotebookApp.default_url = '/edit/civis-notebook-logs.log'
+else:
+    nb_file_path = os.environ.get('NOTEBOOK_FILE_PATH', 'notebook.ipynb').strip('/')
+    notebook_full_path = os.path.join(ROOT_DIR, nb_file_path)
+    c.NotebookApp.default_url = '/notebooks/{}'.format(nb_file_path)
 
-REQUIREMENTS_PATH = os.path.expanduser(os.path.join('~', 'work', 'requirements.txt'))
+    try:
+        # pull .ipynb file from s3 and save preview back if necessary
+        # pull requirements.txt file from s3
+        if not os.path.isfile(notebook_full_path):
+            platform_persistence.initialize_notebook_from_platform(notebook_full_path)
 
-if os.path.isfile(REQUIREMENTS_PATH):
-    platform_persistence.logger.info('installing requirements.txt packages')
-    pip.main(['install', '-r', REQUIREMENTS_PATH])
-    platform_persistence.logger.info('requirements.txt installed')
+        _, preview_url = platform_persistence.get_update_urls()
+        platform_persistence.generate_and_save_preview(preview_url, notebook_full_path)
+
+    except NotebookManagementError as e:
+        platform_persistence.logger.error(str(e))
+        platform_persistence.logger.warn('Killing the notebook process b/c of a startup issue')
+        os.kill(os.getpid(), signal.SIGTERM)
+
+    # Install requirements.txt
+    REQUIREMENTS_PATH = os.path.join(ROOT_DIR, 'requirements.txt')
+    if os.path.isfile(REQUIREMENTS_PATH):
+        platform_persistence.logger.info('installing requirements.txt packages')
+        pip.main(['install', '-r', REQUIREMENTS_PATH])
+        platform_persistence.logger.info('requirements.txt installed')
